requirements.txt
# Core Dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Deep Learning
tensorflow>=2.13.0
keras>=2.13.0

# Clustering
hdbscan>=0.8.33

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Data Processing
joblib>=1.3.0
pyyaml>=6.0

# Logging and Monitoring
tqdm>=4.65.0
python-json-logger>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0

# Medical/Healthcare specific
cryptography>=41.0.0  # For HIPAA compliance encryption
